Back to py tests passing with split file organization
