+++ py/BUILD
@@ -82,7 +82,13 @@ py_library(
 
 py_library(
     name = "runtime",
-    srcs = ["formak/runtime/__init__.py"],
+    srcs = [
+        "formak/runtime/__init__.py",
+        "formak/runtime/common.py",
+        "formak/runtime/managed_filter.py",
+        "formak/runtime/managed_rollback.py",
+        "formak/runtime/storage.py",
+    ],
     imports = ["."],
 )
